Refactor Due date presenter
